Squashed a small bug around postgre decimal values

+++ server/migrations/20220317042155-create-project.js
@@ -18,7 +18,7 @@ module.exports = {
         type: Sequelize.DATE
       },
       goal: {
-        type: Sequelize.DECIMAL
+        type: Sequelize.FLOAT
       },
       campaign: {
         type: Sequelize.TEXT
